requirements.txt
Django==1.8.2
argparse==1.2.1
django-crispy-forms==1.4.0
ipdb==0.8.1
ipython==3.1.0
wsgiref==0.1.2
